+++ src/jvmMain/kotlin/KtUniversalisClient.kt
@@ -101,12 +101,12 @@ actual object KtUniversalisClient {
     suspend fun getMarketBoardCurrentData(
         worldDcRegion: String,
         itemIds: ShortArray,
-        listings: Long? = null,
-        entries: Long? = null,
+        listings: Int? = null,
+        entries: Int? = null,
         noGst: Boolean? = null,
         hq: Boolean? = null,
-        statsWithin: Long? = null,
-        entriesWithin: Long? = null,
+        statsWithin: Int? = null,
+        entriesWithin: Int? = null,
         fields: Array<String>? = null,
     ): MarketBoardCurrentData {
         val marketBoardCurrentData =
